requirements.txt
Django==3.0.6
Pillow==7.1.2
django-ckeditor==5.9.0
easy-thumbnails==2.7
